+++ src/client/theme-default/components/VPSidebarItem.vue
@@ -227,12 +227,13 @@ function onCaretClick() {
 
 .caret-icon {
   font-size: 18px;
+  /*rtl:ignore*/
   transform: rotate(90deg);
   transition: transform 0.25s;
 }
 
 .VPSidebarItem.collapsed .caret-icon {
-  transform: rotate(0);
+  transform: rotate(0)/*rtl:rotate(180deg)*/;
 }
 
 .VPSidebarItem.level-1 .items,