+++ src/node/cli.ts
@@ -1,14 +1,18 @@
 import minimist from 'minimist'
 import c from 'picocolors'
-import { createLogger } from 'vite'
+import { createLogger, type Logger } from 'vite'
 import { build, createServer, serve } from '.'
-import { init } from './init/init'
 import { version } from '../../package.json'
+import { init } from './init/init'
 import { bindShortcuts } from './shortcuts'
+
+if (process.env.DEBUG) {
+  Error.stackTraceLimit = Infinity
+}
 
 const argv: any = minimist(process.argv.slice(2))
 
-const logVersion = (logger = createLogger()) => {
+const logVersion = (logger: Logger) => {
   logger.info(`\n  ${c.green(`${c.bold('vitepress')} v${version}`)}\n`, {
     clear: !logger.hasWarned
   })
@@ -56,9 +60,13 @@ if (!command || command === 'dev') {
   createLogger().info('', { clear: true })
   init(argv.root)
 } else {
-  logVersion()
   if (command === 'build') {
-    build(root, argv).catch((err) => {
+    build(root, {
+      ...argv,
+      onAfterConfigResolve(siteConfig) {
+        logVersion(siteConfig.logger)
+      }
+    }).catch((err) => {
       createLogger().error(
         `${c.red(`build error:`)}\n${err.message}\n${err.stack}`
       )