+++ src/node/config.ts
@@ -1,6 +1,6 @@
 import _debug from 'debug'
 import fs from 'fs-extra'
-import path from 'path'
+import path from 'node:path'
 import c from 'picocolors'
 import {
   createLogger,