+++ src/node/markdownToVue.ts
@@ -2,7 +2,7 @@ import { resolveTitleFromToken } from '@mdit-vue/shared'
 import _debug from 'debug'
 import fs from 'fs-extra'
 import { LRUCache } from 'lru-cache'
-import path from 'path'
+import path from 'node:path'
 import type { SiteConfig } from './config'
 import {
   createMarkdownRenderer,