+++ src/node/plugins/rewritesPlugin.ts
@@ -1,5 +1,5 @@
-import type { Plugin } from 'vite'
 import { compile, match } from 'path-to-regexp'
+import type { Plugin } from 'vite'
 import type { SiteConfig, UserConfig } from '../siteConfig'
 
 export function resolveRewrites(